move scala3Compiler config to compiler module
