+++ compiler/src/dotty/tools/dotc/core/tasty/TastyUnpickler.scala
@@ -5,6 +5,7 @@ package tasty
 import scala.language.unsafeNulls
 
 import dotty.tools.tasty.{TastyFormat, TastyBuffer, TastyReader, TastyHeaderUnpickler, UnpicklerConfig}
+import TastyHeaderUnpickler.TastyVersion
 import TastyFormat.NameTags._, TastyFormat.nameTagToString
 import TastyBuffer.NameRef
 
@@ -24,6 +25,39 @@ object TastyUnpickler {
     def apply(ref: NameRef): TermName = names(ref.index)
     def contents: Iterable[TermName] = names
   }
+
+  trait Scala3CompilerConfig extends UnpicklerConfig:
+    private def asScala3Compiler(version: TastyVersion): String =
+      if (version.major == 28) {
+        // scala 3.x.y series
+        if (version.experimental > 0)
+          // scenario here is someone using 3.4.0 to read 3.4.1-RC1-NIGHTLY, in this case, we should show 3.4 nightly.
+          s"the same nightly or snapshot Scala 3.${version.minor - 1} compiler"
+        else s"a Scala 3.${version.minor}.0 compiler or newer"
+      }
+      else if (version.experimental > 0) "the same Scala compiler" // unknown major version, just say same
+      else "a more recent Scala compiler" // unknown major version, just say later
+
+    /** The description of the upgraded scala compiler that can read the given TASTy version */
+    final def upgradedReaderTool(version: TastyVersion): String = asScala3Compiler(version)
+
+    /** The description of the upgraded scala compiler that can produce the given TASTy version */
+    final def upgradedProducerTool(version: TastyVersion): String = asScala3Compiler(version)
+
+    final def recompileAdditionalInfo: String = """
+      |  Usually this means that the library dependency containing this file should be updated.""".stripMargin
+
+    final def upgradeAdditionalInfo(fileVersion: TastyVersion): String =
+      if (fileVersion.isExperimental && experimentalVersion == 0) {
+        """
+          |  Note that you are using a stable compiler, which can not read experimental TASTy.""".stripMargin
+      }
+      else ""
+  end Scala3CompilerConfig
+
+  /** A config for the TASTy reader of a scala 3 compiler */
+  val scala3CompilerConfig = new Scala3CompilerConfig with UnpicklerConfig.DefaultTastyVersion {}
+
 }
 
 import TastyUnpickler._
@@ -88,7 +122,7 @@ class TastyUnpickler(reader: TastyReader) {
     result
   }
 
-  new TastyHeaderUnpickler(UnpicklerConfig.scala3Compiler, reader).readHeader()
+  new TastyHeaderUnpickler(scala3CompilerConfig, reader).readHeader()
 
   locally {
     until(readEnd()) { nameAtRef.add(readNameContents()) }
